Move all sql code to fili core
