+++ fili-core/src/main/java/com/yahoo/bard/webservice/sql/database/Database.java
@@ -1,6 +1,6 @@
 // Copyright 2016 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.test;
+package com.yahoo.bard.webservice.sql.database;
 
 import com.yahoo.bard.webservice.sql.helper.TimestampUtils;
