+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/sql/evaluator/FilterEvaluatorSpec.groovy
@@ -2,18 +2,18 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.sql.evaluator
 
-import static com.yahoo.bard.webservice.sql.builders.Filters.and
-import static com.yahoo.bard.webservice.sql.builders.Filters.not
-import static com.yahoo.bard.webservice.sql.builders.Filters.or
-import static com.yahoo.bard.webservice.sql.builders.Filters.search
-import static com.yahoo.bard.webservice.sql.database.Database.ID
-import static com.yahoo.bard.webservice.sql.database.Database.IS_NEW
-import static com.yahoo.bard.webservice.sql.database.Database.IS_ROBOT
-import static com.yahoo.bard.webservice.sql.database.Database.METRO_CODE
-import static com.yahoo.bard.webservice.sql.database.Database.WIKITICKER
+import static com.yahoo.bard.webservice.sql.helper.Filters.and
+import static com.yahoo.bard.webservice.sql.helper.Filters.not
+import static com.yahoo.bard.webservice.sql.helper.Filters.or
+import static com.yahoo.bard.webservice.sql.helper.Filters.search
+import static com.yahoo.bard.webservice.sql.helper.SimpleDruidQueryBuilder.ID
+import static com.yahoo.bard.webservice.sql.helper.SimpleDruidQueryBuilder.IS_NEW
+import static com.yahoo.bard.webservice.sql.helper.SimpleDruidQueryBuilder.IS_ROBOT
+import static com.yahoo.bard.webservice.sql.helper.SimpleDruidQueryBuilder.METRO_CODE
+import static com.yahoo.bard.webservice.sql.helper.SimpleDruidQueryBuilder.WIKITICKER
 
-import com.yahoo.bard.webservice.sql.database.Database
 import com.yahoo.bard.webservice.sql.helper.CalciteHelper
+import com.yahoo.bard.webservice.sql.database.Database
 
 import org.apache.calcite.rel.rel2sql.RelToSqlConverter
 import org.apache.calcite.sql.SqlDialect
@@ -33,6 +33,7 @@ class FilterEvaluatorSpec extends Specification {
     def "GetDimensionNames expecting #dimensions"() {
         setup:
         RelBuilder builder = CalciteHelper.getBuilder(Database.getDataSource())
+        FilterEvaluator filterEvaluator = new FilterEvaluator()
         builder.scan(WIKITICKER)
         def rexNodes = dimensions.stream()
                 .map { builder.field(it) }
