+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/sql/helper/Havings.groovy
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.helper
+package com.yahoo.bard.webservice.sql.helper
 
 import com.yahoo.bard.webservice.druid.model.having.AndHaving
 import com.yahoo.bard.webservice.druid.model.having.Having
