+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/sql/helper/SimpleDruidQueryBuilder.groovy
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.helper
+package com.yahoo.bard.webservice.sql.helper
 
 import com.yahoo.bard.webservice.data.config.dimension.DefaultDimensionField
 import com.yahoo.bard.webservice.data.config.dimension.DefaultKeyValueStoreDimensionConfig
