--- fili-sql-backend/src/main/resources/applicationConfig.properties
@@ -1,42 +1,0 @@
-# Copyright 2016 Yahoo Inc.
-# Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-
-# Application configuration must specify a package name.  This prefix applies to all config properties.
-package_name = bard
-
-bard__version = Local Mode
-bard__resource_binder=com.yahoo.wiki.webservice.application.GenericBinderFactory
-
-bard__non_ui_druid_broker=http://localhost:8082/druid/v2
-bard__ui_druid_broker=http://localhost:8082/druid/v2
-bard__druid_coord=http://localhost:8081/druid/coordinator/v1
-bard__fili_port=9998
-
-# Use memory for the default dimension backing store
-bard__dimension_backend=memory
-
-# Data Cache
-bard__druid_cache_enabled = false
-bard__druid_cache_v2_enabled = false
-
-# Flag to enable usage of metadata supplied by the druid coordinator
-# It requires coordinator URL (bard__druid_coord) to be set
-bard__druid_coordinator_metadata_enabled = true
-
-# Lucene index files path
-bard__lucene_index_path=/home/y/var/
-
-# maximum number of results to display without any filters - used for /dim/values endpoint
-bard__max_results_without_filters=10000
-
-# Default number of records per-page
-bard__default_per_page=10000
-
-# Enable permissive partial data, marking things as partial only when no metrics are available rather then when only one is not available
-bard__permissive_column_availability_enabled=false
-
-# Enable TopN optimization
-bard__top_n_enabled=true
-
-# Flag to turn on case sensitive keys in keyvalue store
-bard__case_sensitive_keys_enabled = false
