--- fili-sql-backend/src/main/resources/userConfig.properties.sample
@@ -1,5 +1,0 @@
-# Copyright 2016 Yahoo Inc.
-# Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-
-# Overrides the default dimension backend.
-fili__dimension_backend = memory
